# doc/api.yaml
openapi: 3.0.3
info:
  title: HIU-PLATEFORM-API
  description: |-
    it's a cybersec challenge evaluation plateform
  version: 3.2.0
servers:
  - url: http://localhost:8080
paths:
...
  /user:
    post:
      tags:
        - User
      summary: Add a new user
      description: Add a new user
      operationId: addUser
      requestBody:
        description: User object to be added
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        400:
          $ref: '#/components/responses/400'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        429:
          $ref: '#/components/responses/429'
        500:
          $ref: '#/components/responses/500'
    get:
      tags:
        - User
      summary: Retrieve users
      description: Retrieve a list of users for simple user
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        400:
          $ref: '#/components/responses/400'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        429:
          $ref: '#/components/responses/429'
        500:
          $ref: '#/components/responses/500'
  /user/admin:
    get:
      tags:
        - User
      summary: Retrieve users with administrator permission
      description: Retrieve a list of users by administrator acces
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        400:
          $ref: '#/components/responses/400'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        429:
          $ref: '#/components/responses/429'
        500:
          $ref: '#/components/responses/500'
  /user/{id}:
    get:
      tags:
        - User
      summary: Retrieve user
      description: find an user
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the user to find
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        400:
          $ref: '#/components/responses/400'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        429:
          $ref: '#/components/responses/429'
        500:
          $ref: '#/components/responses/500'
    put:
      tags:
        - User
      security:
        - jwt-token: [ ]
      summary: Update an user by ID
      description: Update an user by ID
      operationId: updateUser
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the user to be updated
          schema:
            type: string
      requestBody:
        description: User object to update
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        '200':
          description: Operation successful, the user has been updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
            application/xml:
              schema:
                $ref: "#/components/schemas/User"
        400:
          $ref: '#/components/responses/400'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        429:
          $ref: '#/components/responses/429'
        500:
          $ref: '#/components/responses/500'
    delete:
      tags:
        - User
      security:
        - jwt-token: [ ]
      summary: Delete a user by ID
      description: Delete a user by ID
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the user to be deleted
          schema:
            type: string
      responses:
        '200':
          description: Operation successful, the user has been deleted
        400:
          $ref: '#/components/responses/400'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        429:
          $ref: '#/components/responses/429'
        500:
          $ref: '#/components/responses/500'
...
  /response:
    post:
      tags:
        - Response
      summary: Add a new response
      description: Add a new response
      operationId: addResponse
      requestBody:
        description: Response object to be added
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Response'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
        400:
          $ref: '#/components/responses/400'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        429:
          $ref: '#/components/responses/429'
        500:
          $ref: '#/components/responses/500'
    get:
      tags:
        - Response
      summary: Retrieve responses
      description: Retrieve a list of responses
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Response'
        400:
          $ref: '#/components/responses/400'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        429:
          $ref: '#/components/responses/429'
        500:
          $ref: '#/components/responses/500'
  /response/{id}:
    get:
      tags:
        - Response
      summary: Retrieve team's responses
      description: Retrieve a list of team's responses
      parameters:
        - in: path
          name: id
          description: team's id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Response'
        400:
          $ref: '#/components/responses/400'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        429:
          $ref: '#/components/responses/429'
        500:
          $ref: '#/components/responses/500'
  /response/team-by-subject:
    get:
      tags:
        - Response
      summary: Get team list group by subject they answered
      description: Get team list group by subject they answered
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TeamListGroupBySubject'
        400:
          $ref: '#/components/responses/400'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        429:
          $ref: '#/components/responses/429'
        500:
          $ref: '#/components/responses/500'
...
components:
  schemas:
    User:
      type: object
      required:
        - id
        - user_name
        - email
        - password
        - role
      properties:
        id:
          type: integer
          format: int64
        user_name:
          type: string
        email:
          type: string
        password:
          type: string
        profile_picture:
          type: string
          format: byte
        id_team:
          type: string
          example:
            $ref: "#/components/schemas/Challengers_team"
        role:
          type: string
          description: role of the user by default user
          example: "USER, ADMIN"
    University:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
    Discipline:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
    Subject:
      type: object
      required:
        - id
        - title
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        max_score:
          type: number
          format: float
    Challengers_team:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        slogan:
          type: string
        id_university:
          type: string
          example:
            $ref: "#/components/schemas/University"
        id_discipline:
          type: string
          example:
            $ref: "#/components/schemas/Discipline"
    Response:
      type: object
      required:
        - id
        - response
      properties:
        id:
          type: string
        response:
          type: string
        commentary:
          type: string
        id_challenger_team:
          type: string
          example:
            $ref: "#/components/schemas/Challengers_team"
        id_subject:
          type: string
          example:
            $ref: "#/components/schemas/Subject"
    TeamListGroupBySubject:
      type: object
      required:
        - id_subject
        - id_team
      properties:
        id_subject:
          type: string
          $ref: "#/components/schemas/Subject"
        id_team:
          type: string
          $ref: "#/components/schemas/Challengers_team"
...
    Excepetion:
      type: object
      properties:
        type:
          type: string
        message:
          type: string
    BadRequestException:
      allOf:
        - $ref: '#/components/schemas/Excepetion'
      example:
        type: BadRequestException
        message: Bad request
    NoAuthorizedException:
      allOf:
        - $ref: '#/components/schemas/Excepetion'
      example:
        type: NotAuthorizedException
        message: Not authorized
    RessourceNotFoundException:
      allOf:
        - $ref: '#/components/schemas/Excepetion'
      example:
        type: RessourceNOtFoundException
        message: Ressource of type <T> identified by <I> not found
    TooManyRequestsException:
      allOf:
        - $ref: '#/components/schemas/Excepetion'
      example:
        type: TooManyRequestsException
        message: Too many requests
    InternalServerException:
      allOf:
        - $ref: '#/components/schemas/Excepetion'
      example:
        type: InternalServerException
        message: Unexpected error
  responses:
    '400':
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BadRequestException'
    '403':
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/NoAuthorizedException'
    '404':
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RessourceNotFoundException'
    '429':
      description: Too many requests to the API
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/TooManyRequestsException'
    '500':
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/InternalServerException'