# src/main/java/com/hui/plateform/hiuplateformeapi/entity/User.java
package com.hui.plateform.hiuplateformeapi.entity;

import jakarta.persistence.*;
import lombok.*;
import org.springframework.security.core.GrantedAuthority;
import org.springframework.security.core.userdetails.UserDetails;

import java.util.Collection;

import static jakarta.persistence.GenerationType.IDENTITY;

@Entity
@Builder
@AllArgsConstructor
@NoArgsConstructor
@Data
@Table(name = "user_table")
public class  User implements UserDetails {
    @Id
    @Column(name = "id")
    @GeneratedValue(strategy = IDENTITY)
    private String id;

    @Column(name = "email")
    private String email;

    @Column(name = "username")
    private String username;

    @Column(name = "password")
    private String password;

    @Column(name = "profile_picture")
    private String profilePicture;

    @Column(name = "id_challenger")
    private String idChallenger;

    @Enumerated(value = EnumType.STRING)
    @Column(name = "role")
    Role role;
    @Override
    public Collection<? extends GrantedAuthority> getAuthorities() {
        return null;
    }

    @Override
    public String getPassword() {
        return password;
    }

    @Override
    public boolean isAccountNonExpired() {
        return true;
    }

    @Override
    public boolean isAccountNonLocked() {
        return true;
    }

    @Override
    public boolean isCredentialsNonExpired() {
        return true;
    }

    @Override
    public String getUsername(){
        return username;
    }
    @Override
    public boolean isEnabled() {
        return true;
    }
}